+++ redline/__init__.py
@@ -1,5 +1,5 @@
-from .watch import watch
+from .watch import record, watch
 
-__all__ = ["__version__", "watch"]
+__all__ = ["__version__", "record", "watch"]
 
 __version__ = "0.1.0"